refactor: change hook location
